Fix Issue w/ Screenshot of Strava Relative Effort
Clicking the relative effort tab loads an SVG. The process needs to wait
until it's visible because the tab is on the screen when it loads.

For some reason moving this tab to the last one made it work. Otherwise,
Puppeteer waits on the element to be visible forever. I'm not sure why
this is because the SVG loads when it is clicked, not when the page
loads.